Backend/Ceema/requirements.txt
Django==5.2.12
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
